17-Multi-agent-memory/requirements.txt
google-adk[database]==1.4.2
yfinance==0.2.56
psutil==5.9.5
litellm==1.66.3
google-genai==1.21.1
python-dotenv==1.1.0
deprecated==1.2.14
uvicorn
psycopg2-binary
sqlalchemy 